Hack because of a teacher
Had a teacher that put in zeros for all the assignments; now they are all marked as graded and don't sync. Had to revert a change to look at workflow status and instead just look a t submission attempts. This will break cases where teacher enters a grade manually without submissions; I'll have to change the logic here to do a couple of things instead of one big OR but need it working for now.